requirements.txt
fastapi
uvicorn
python-multipart
transformers
Pillow
torch
tf-keras    